Created page where users can define their favorite foods

+++ nutrition-tracker-frontend/src/features/favorites/favoriteApi.ts
@@ -1,36 +1,36 @@
 import { supabase } from '@/lib/supabase'
-import { Database } from '@/types/database.types'
-
-type Favorite = Database['public']['Tables']['favorite_foods']['Row']
-type FavoriteInsert = Database['public']['Tables']['favorite_foods']['Insert']
 
 // mirrors SELECT statement policy
-export const getFavorites = async (): Promise<Favorite[]> => {
+export const getFavoriteFoods = async (userId: string) => {
   const { data, error } = await supabase
     .from('favorite_foods')
-    .select(`*, foods(*)`)
+    .select(`
+      food_id,
+      added_at,
+      foods(*)`)
+    .eq('user_id', userId)
+    .order('added_at', {ascending: false})
 
   if (error) throw error
   return data
 }
 
 // mirrors INSERT statement policy
-export const addFavorite = async (fav: FavoriteInsert) => {
+export const addFavorite = async (userId: string, foodId: number) => {
   const { data, error } = await supabase
     .from('favorite_foods')
-    .insert(fav)
+    .insert({ user_id: userId, food_id: foodId})
 
   if (error) throw error
-  return data
 }
 
 // mirrors DELETE statement policy
-export const deleteFavorite = async (foodId: number) => {
-  const { data, error } = await supabase
+export const removeFavorite = async (userId: string, foodId: number) => {
+  const { error } = await supabase
     .from('favorite_foods')
     .delete()
+    .eq('user_id', userId)
     .eq('food_id', foodId)
   
   if (error) throw error
-  return data
 }
